add useShiftTicketMetrics, reactor TurnProvider logic, delete useShiftTickets.
