--- client/src/hooks/turn/useShiftTickets.js
@@ -1,90 +1,0 @@
-"use client";
-import { useEffect, useState } from "react";
-
-import { useTranslations } from "next-intl";
-
-import {
-  getTickets,
-  getPayments,
-  getPaymentMethods,
-  getPaymentByTicketId,
-} from "@/services/ticketsService";
-
-export function useShiftTickets(shiftData) {
-  const shiftTranslations = useTranslations("shifts");
-  const [totalBalance, setTotalBalance] = useState(0);
-  const [totalTickets, setTotalTickets] = useState(0);
-  const [byPaymentMethod, setByPaymentMethod] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [breakdownLoading, setBreakdownLoading] = useState(false);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    if (!shiftData?.shiftDate || !shiftData?.startTime) return;
-
-    let cancelled = false;
-
-    async function fetchBreakdown(shiftTickets) {
-      if (!cancelled) setBreakdownLoading(true);
-      try {
-        const [payments, methods] = await Promise.all([
-          getPayments(),
-          getPaymentMethods(),
-        ]);
-
-        const methodTotals = {};
-        for (const ticket of shiftTickets) {
-          const ticketPayments = await getPaymentByTicketId(ticket.id);
-          if (!ticketPayments?.length) continue;
-
-          const payment = payments.find((payment) => payment.id === ticketPayments[0].paymentId);
-          if (!payment) continue;
-
-          const method = methods.find((method) => method.id === payment.methodId);
-          const methodName = method?.name ?? shiftTranslations("other");
-
-          methodTotals[methodName] = (methodTotals[methodName] ?? 0) + ticket.totalAmount;
-        }
-
-        if (!cancelled) {
-          setByPaymentMethod(
-            Object.entries(methodTotals).map(([name, total]) => ({ name, total })),
-          );
-        }
-      } finally {
-        if (!cancelled) setBreakdownLoading(false);
-      }
-    }
-
-    async function fetchData() {
-      setLoading(true);
-      setError(null);
-      try {
-        const shiftStartMs = new Date(
-          `${shiftData.shiftDate}T${shiftData.startTime}`,
-        ).getTime();
-
-        const tickets = await getTickets();
-        const shiftTickets = tickets.filter(
-          (ticket) => new Date(`${ticket.ticketDate.replace(" ", "T")}Z`).getTime() >= shiftStartMs,
-        );
-
-        if (cancelled) return;
-
-        setTotalBalance(shiftTickets.reduce((runningTotal, ticket) => runningTotal + ticket.totalAmount, 0));
-        setTotalTickets(shiftTickets.length);
-
-        fetchBreakdown(shiftTickets).catch(() => {});
-      } catch (err) {
-        if (!cancelled) setError(err?.message || shiftTranslations("loadError"));
-      } finally {
-        if (!cancelled) setLoading(false);
-      }
-    }
-
-    fetchData();
-    return () => { cancelled = true; };
-  }, [shiftData?.shiftDate, shiftData?.startTime, shiftTranslations]);
-
-  return { totalBalance, totalTickets, byPaymentMethod, loading, breakdownLoading, error };
-}
