+++ client/src/providers/turn/TurnProvider.jsx
@@ -1,21 +1,13 @@
 "use client";
 import { createContext, useCallback, useEffect, useMemo, useState } from "react";
 
-import { useTranslations } from "next-intl";
-
 import { useAuth } from "@/hooks/auth/useAuth";
+import { useShiftTicketMetrics } from "@/hooks/turn/useShiftTicketMetrics";
 import { getTurnOpen, openTurn, closeTurn } from "@/services/shiftsService";
-import {
-  getTickets,
-  getPayments,
-  getPaymentMethods,
-  getPaymentByTicketId,
-} from "@/services/ticketsService";
 
 export const TurnContext = createContext();
 
 export function TurnProvider({ children }) {
-  const shiftTranslations = useTranslations("shifts");
   const { user } = useAuth();
 
   const [openTurnId, setOpenTurnId] = useState(null);
@@ -23,11 +15,15 @@ export function TurnProvider({ children }) {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
 
-  const [totalBalance, setTotalBalance] = useState(0);
-  const [totalTickets, setTotalTickets] = useState(0);
-  const [byPaymentMethod, setByPaymentMethod] = useState([]);
-  const [ticketsLoading, setTicketsLoading] = useState(false);
-  const [breakdownLoading, setBreakdownLoading] = useState(false);
+  const {
+    totalBalance,
+    totalTickets,
+    byPaymentMethod,
+    ticketsLoading,
+    breakdownLoading,
+    refresh: refreshShiftTickets,
+    reset: resetTicketMetrics,
+  } = useShiftTicketMetrics(openShiftData);
 
   const loadOpenTurn = useCallback(async () => {
     try {
@@ -36,8 +32,8 @@ export function TurnProvider({ children }) {
       setOpenTurnId(shift?.id ?? null);
       setOpenShiftData(shift);
       return shift?.id ?? null;
-    } catch (err) {
-      setError(err?.message || "Error al obtener turno");
+    } catch (caughtError) {
+      setError(caughtError?.message || "Error al obtener turno");
       setOpenTurnId(null);
       setOpenShiftData(null);
       return null;
@@ -49,72 +45,6 @@ export function TurnProvider({ children }) {
   useEffect(() => {
     loadOpenTurn();
   }, [loadOpenTurn, user?.id]);
-
-  const fetchTicketBreakdown = useCallback(async (shiftTickets) => {
-    setBreakdownLoading(true);
-    try {
-      const [payments, methods] = await Promise.all([
-        getPayments(),
-        getPaymentMethods(),
-      ]);
-
-      const methodTotals = {};
-      for (const ticket of shiftTickets) {
-        const ticketPayments = await getPaymentByTicketId(ticket.id);
-        if (!ticketPayments?.length) continue;
-
-        const payment = payments.find((payment) => payment.id === ticketPayments[0].paymentId);
-        if (!payment) continue;
-
-        const method = methods.find((method) => method.id === payment.methodId);
-        const methodName = method?.name ?? shiftTranslations("other");
-
-        methodTotals[methodName] = (methodTotals[methodName] ?? 0) + ticket.totalAmount;
-      }
-
-      setByPaymentMethod(
-        Object.entries(methodTotals).map(([name, total]) => ({ name, total })),
-      );
-    } finally {
-      setBreakdownLoading(false);
-    }
-  }, [shiftTranslations]);
-
-  const fetchShiftTickets = useCallback(async () => {
-    if (!openShiftData?.shiftDate || !openShiftData?.startTime) return;
-
-    setTicketsLoading(true);
-    try {
-      const shiftStartMs = new Date(
-        `${openShiftData.shiftDate}T${openShiftData.startTime}`,
-      ).getTime();
-
-      const tickets = await getTickets();
-      const shiftTickets = tickets.filter(
-        (ticket) => new Date(`${ticket.ticketDate.replace(" ", "T")}Z`).getTime() >= shiftStartMs,
-      );
-
-      setTotalBalance(shiftTickets.reduce((runningTotal, ticket) => runningTotal + ticket.totalAmount, 0));
-      setTotalTickets(shiftTickets.length);
-
-      fetchTicketBreakdown(shiftTickets).catch(() => {});
-    } catch {
-    } finally {
-      setTicketsLoading(false);
-    }
-  }, [openShiftData?.shiftDate, openShiftData?.startTime, fetchTicketBreakdown]);
-
-  useEffect(() => {
-    fetchShiftTickets();
-  }, [fetchShiftTickets]);
-
-  const resetTicketMetrics = useCallback(() => {
-    setTotalBalance(0);
-    setTotalTickets(0);
-    setByPaymentMethod([]);
-    setTicketsLoading(false);
-    setBreakdownLoading(false);
-  }, []);
 
   const updateTurn = useCallback((newTurnId) => {
     setOpenTurnId(newTurnId);
@@ -162,13 +92,13 @@ export function TurnProvider({ children }) {
       byPaymentMethod,
       ticketsLoading,
       breakdownLoading,
-      refreshShiftTickets: fetchShiftTickets,
+      refreshShiftTickets,
     }),
     [
       openTurnId, openShiftData, loading, error,
       updateTurn, refreshTurn, loadOpenTurn, openShift, closeShift,
       totalBalance, totalTickets, byPaymentMethod, ticketsLoading, breakdownLoading,
-      fetchShiftTickets,
+      refreshShiftTickets,
     ],
   );
